requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2
django-quill-editor==0.1.40
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.17.0
packaging==21.3
Pillow==9.1.1
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
